requirements.txt
aiofiles
aiohttp
future
gTTS
googletrans
bs4
pytz
requests
TgCrypto
pyromod==1.1.0
ffmpeg-python
pymongo
Flask
gpytranslate
Telethon==1.24.0
cryptg
typing
Python-IO
Pillow==9.5.0
telegraph
motor
SQLAlchemy
opencv-python-headless
opencv-python
prettytable
svglib
reportlab
uwuify
python-dotenv
humanize
flask_restful
googlesearch-python
gitpython
pymediainfo
prettytable
httpx==0.27.1
speedtest-cli
numpy
validators
emoji
youtube_search_python
youtube_search
rich
wget
hachoir
pykeyboard
python-arq
search-engine-parser
uvloop
feedparser
fuzzysearch
img2pdf
pytz
psutil
google-search-results
uvicorn
fastapi[all]
typing-extensions
uvicorn[standard]
pydantic
g4f
aiocache
aiogram
aiosqlite
apscheduler
arrow
GitPython
qrcode
pygments
environs
curl_cffi
nest_asyncio
cohere
icrawler
unidecode
git-python
pySmartDL
selenium
PyYAML==6.0.2
kurigram
akenoai[fast]
git+https://github.com/johnwmillr/LyricsGenius
yt-dlp==2024.8.6
